+++ server/frontend/src/lib/panes/PaneBrowserSample/index.js
@@ -1,4 +1,5 @@
 export { default as PaneBrowserSample } from './PaneBrowserSample.vue'
 export { default as BatchContextMenu } from './BatchContextMenu.vue'
 export { default as DatasetContextMenu } from './DatasetContextMenu.vue'
+export { default as DialogPasteIntoNew } from './DialogPasteIntoNew.vue'
 export * from './stores'